use i32 instead of i64 for paged_attention

+++ src/cpp/src/utils.cpp
@@ -267,10 +267,10 @@ void apply_gather_before_matmul_transformation(std::shared_ptr<ov::Model> model)
     std::tie(matmul, slice_gather_dim) = find_llm_matmul(model);
 
     if (matmul && matmul->input(0).get_partial_shape().rank().get_length() == 3) {
-        auto indices = std::make_shared<ov::op::v0::Parameter>(ov::element::i64, ov::PartialShape{-1});
+        auto indices = std::make_shared<ov::op::v0::Parameter>(ov::element::i32, ov::PartialShape{-1});
         indices->set_friendly_name("sampled_tokens_indices");
         indices->output(0).get_tensor().set_names({"sampled_tokens_indices"});
-        auto axis = std::make_shared<ov::op::v0::Constant>(ov::element::i64, ov::Shape{1}, std::vector<int64_t>{slice_gather_dim});
+        auto axis = std::make_shared<ov::op::v0::Constant>(ov::element::i32, ov::Shape{1}, std::vector<int32_t>{slice_gather_dim});
         auto gather = std::make_shared<ov::op::v8::Gather>(matmul->input_value(0), indices, axis);
         matmul->input(0).replace_source_output(gather);
         model->add_parameters({indices});
